learning/pyproject.toml
[tool.poetry]
name = "peano"
version = "0.1.0"
description = "An environment for learning formal mathematical reasoning from exploration"
authors = ["Gabriel Poesia <me@gpoesia.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.11.0"
hydra-core = "^1.1.2"
wandb = "^0.12.14"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
